+++ cardinal-components-base/src/main/java/dev/onyxstudios/cca/api/v3/component/ServerTickingComponent.java
@@ -20,32 +20,19 @@
  * OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE
  * OR OTHER DEALINGS IN THE SOFTWARE.
  */
-package dev.onyxstudios.cca.api.v3.chunk;
+package dev.onyxstudios.cca.api.v3.component;
 
-import nerdhub.cardinal.components.api.component.Component;
-import net.minecraft.world.chunk.Chunk;
-import org.jetbrains.annotations.Contract;
+import net.minecraft.util.Identifier;
+import org.jetbrains.annotations.ApiStatus;
 
 /**
- * A component factory for {@linkplain Chunk chunks}.
+ * A component that gets ticked alongside the provider it is attached to.
  *
- * <p>When invoked, the factory must return a {@link Component} of the right type.
- *
- * @since 2.4.0
+ * <p>This interface must be visible at factory registration time - which means the class implementing it
+ * must either be the parameter to {@link ComponentRegistryV3#getOrCreate(Identifier, Class)} or declared explicitly
+ * using a dedicated method on the factory registry.
  */
-@FunctionalInterface
-public interface ChunkComponentFactory<C extends Component> {
-    /**
-     * Initialize components for the given chunk.
-     *
-     * <p>The component returned by this method will be available
-     * on the chunk as soon as all component factories have been invoked.
-     *
-     * @param chunk      the chunk being constructed
-     * @implNote Because this method is called for each chunk creation, implementations
-     * should avoid side effects and keep costly computations at a minimum. Lazy initialization
-     * should be considered for components that are costly to initialize.
-     */
-    @Contract(pure = true)
-    C createForChunk(Chunk chunk);
+@ApiStatus.Experimental
+public interface ServerTickingComponent extends ComponentV3 {
+    void tick();
 }