+++ src/testmod/java/dev/onyxstudios/componenttest/vita/PlayerVita.java
@@ -23,7 +23,9 @@
 package dev.onyxstudios.componenttest.vita;
 
 import dev.onyxstudios.cca.api.v3.component.AutoSyncedComponent;
+import dev.onyxstudios.cca.api.v3.component.ServerTickingComponent;
 import dev.onyxstudios.cca.api.v3.entity.PlayerComponent;
+import dev.onyxstudios.componenttest.CardinalComponentsTest;
 import dev.onyxstudios.componenttest.TestComponents;
 import net.minecraft.client.MinecraftClient;
 import net.minecraft.entity.player.PlayerEntity;
@@ -34,7 +36,7 @@ import net.minecraft.server.network.ServerPlayerEntity;
 /**
  * A Vita component attached to players, and automatically synchronized with their owner
  */
-public class PlayerVita extends EntityVita implements AutoSyncedComponent, PlayerComponent<BaseVita> {
+public class PlayerVita extends EntityVita implements AutoSyncedComponent, ServerTickingComponent, PlayerComponent<BaseVita> {
     public static final int INCREASE_VITA = 0b10;
     public static final int DECREASE_VITA = 0b100;
 
@@ -48,6 +50,13 @@ public class PlayerVita extends EntityVita implements AutoSyncedComponent, Playe
             boolean increase = value > this.vitality;
             super.setVitality(value);
             TestComponents.VITA.sync(this.owner, increase ? INCREASE_VITA : DECREASE_VITA);
+        }
+    }
+
+    @Override
+    public void tick() {
+        if (this.owner.age % 1200 == 0) {
+            CardinalComponentsTest.LOGGER.info("{} is still alive", this.owner);
         }
     }